Fixed more problems with the format strings (Reported by Fritz Mehner).
The %F placeholder was not being translated to "STDIN" when the input
comes from STDIN.  This was causing a warning from File::Basename
because the filename was an empty string.

The %r placeholder would sometimes evaluate to an empty string.  This
was because the document (and its elements) were garbage collected
before the placehold was evaluated.  This is just like the problem
with the %f placeholder that Chris fixed.




git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/Perl-Critic@1235 7ac2fd03-a204-0410-b7b1-b9f6acf132f9