Merge r3284 to the backlog branch.
T 36081: RegularExpressions::ProhibitCaptureWithoutTest flags cases
when $1, etc. clearly are set. When backing through the code looking
for a conditional statement, instead of just failing if a match or
substitute is found, we pass if the match (etc) is followed by 'or'
and 'die', 'croak', or 'confess'. Configuration option 'exception
source' allows the user to add other exception-throwing routines.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/branches/Perl-Critic-backlog@3510 7ac2fd03-a204-0410-b7b1-b9f6acf132f9