Merge r3288 to the backlog branch.
ProhibitVersionStrings versus comments within a "use". Instead of just
a regex on the contents of the statement, attempted to actually
identify the version string. This is made more complicated by the fact
that PPI does not correctly parse version strings beginning with a
"v". The .run file was updated with a test like the example provided
in the RT ticket.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/branches/Perl-Critic-backlog@3513 7ac2fd03-a204-0410-b7b1-b9f6acf132f9