Merge r3448, r3615, r3616, and r3649 to the backlog branch.
Assign blame to correct person

Since PPI 1.206 still mistakes constructors for blocks in some (though
fewer) cases, comment code which is hacked to work around this, and
add a test to see when the problem is finally fixed.

Correct name of routine to be fixed when test passes.

Join broken NAME line.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/branches/Perl-Critic-backlog@3659 7ac2fd03-a204-0410-b7b1-b9f6acf132f9