Convert Perl::Critic from using Regexp::Parser for regular expression analysis to using PPIx::Regxp. Besides the P::C::P::RegularExpressions modules, collateral changes were made in Perl::Critic (to flush the cached regular expression parses after criticism is complete).
The new parser introduced self-compliance problems in:
 xt/author/93_version.t
 lib/Perl/Critic/Document.pm
 lib/Perl/Critic/Policy/CodeLayout/RequireConsistentNewlines.pm
 lib/Perl/Critic/Policy/Modules/ProhibitEvilModules.pm
which are addressed in this commit.

Obviously this commit adds PPIx::Regexp as an optional module. But
Regexp::Parser is _not_ dropped, because the relevant subroutines
in P::C::U::PPIRegexp (parse_regexp and ppiify) need to be put
through a deprecation cycle.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3704 7ac2fd03-a204-0410-b7b1-b9f6acf132f9