Modify the annotation parsing so that things like   my $foo =     '$bar'; ## no critic (RequireInterpolationOfMetachars)
are parsed as line annotations rather than block annotations. Test
t/03_annotations.t was added to test this, since I was too lazy to
recode the tests into t/03_pragmas.t.

The statement in P::C::P::Variables::ProhibitPunctuationVars that was
modified (strangely) to work around the old behavior has been put back
the way it was, and the prolix comment on the strange coding removed.



git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3744 7ac2fd03-a204-0410-b7b1-b9f6acf132f9