No longer generating the PolicySummary when running the "test" action. This was causing all the build failures in the last release, because the PolicySummary had already been generated and had read-only permissions when a user receives the distro.  So running "./Build test" would fail because it would try to regenerate the read-only file.
I also noticed that we're deleting and regenerating the MANIFEST
every time that you "./Build test".  This seems like a convenience
for us authors, but for end-users, it seems like awfully weird behavior.
So I took that out as well.

So now, the PolicySummary and MANIFEST are generated only when running
authortest, which always happens before making a distro.  If an author
wants to run some other tests that require the MANIFEST, then it is up
to them to run the "manifest" target.

At least, this is what makes sense to me.  I know Elliot has strong
opionions about this, and went to great lengths to configure this 
build arrangement.  So he may think differently.



git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3761 7ac2fd03-a204-0410-b7b1-b9f6acf132f9