RT #54524: False Positive: ValuesAndExpressions::ProhibitMismatchedOperators with 'foo' x 15 . 'bar'
Brute force fix. Just before returning the violation, check to see if
we're munching on a string operator preceded by a number. If so, check
to see if the thing before the number is an 'x' operator, and if so
simply return. Otherwise fall through.

A test for this was added to the .run file.

No attempt was made to figure out what _kind_ of 'x' operator we had,
since the string operator we're analyzing provides scalar context, so
in something like ( 1, 2, 3 ) x 4 . 'bar' the list gets flattened to its
last element, so we get '3333bar'.

I have not addressed the question of whether 3 x 4 should be a violation
of this policy.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3767 7ac2fd03-a204-0410-b7b1-b9f6acf132f9