RT 55604: Variables::ProhibitPunctuationVars gives false-positive on qr// regexp's ending in '$'
The problem here was that the code was analyzing the entire interpolated
string, including the delimiters, so qr/foo$/ looked like a use of $/.

The proposed fix uses string() to extract the contents of the string if
this method is available (the PPI::Token::Quote objects). If not (the
PPI::Token::QuoteLike objects) it calls content(), but then uses a regex
for the left-hand end of the string, and a chop() for the right-hand
end.

A test for this was added to the .run file also.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3791 7ac2fd03-a204-0410-b7b1-b9f6acf132f9