RT #56619: "die" used instead of "croak" (possibly false positive)
Allow 'warn' or 'die' in name space main::, provided it is not in a
subroutine. This implementation per the BUGS section of the
P::C::P::ErrorHandling::RequireCarping POD (except, of course that it no
longer says that).

Some tests added to RequireCarping.run to exercise the new code, and
(glory be!) Changes updated to say in general what was done.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3799 7ac2fd03-a204-0410-b7b1-b9f6acf132f9