RT #56619 redux
After discussion on the mailing list, added option
allow_in_main_unless_in_subroutine to
P::C::P::ErrorHandling::RequireCarping to control whether 'warn' or
'die' are allowed unconditionally in main:: (unless in a subroutine).
This option is false by default.

The configuration documentation was updated not only to cover this, but
to try to make explicit its interaction with
allow_messages_ending_in_newlines (it's ||, not &&).

The original work was committed as revision 3799.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3801 7ac2fd03-a204-0410-b7b1-b9f6acf132f9