49501 	From Debian: RequirePackageMatchesPodName: doesn't use nearest package
Allow any package (rather than nearest) to match the pod NAME. "Nearest"
was not used because of the practice of putting all pod at the end. The
policy still only checks the first NAME section, if more than one
exists, but since I know of no examples of this, I figure it can wait.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3806 7ac2fd03-a204-0410-b7b1-b9f6acf132f9