RT#61071: OptionsProcessor vs perl 5.8.3 Exporter
Updated inc/Perl/Critic/BuildUtilities.pm to require Exporter version
5.63, to get a version that understands export tags in positions other
than leading ones. Another solution was to move all the tags to the
leading position, but I thought it would be difficult to ensure that the
code stayed that way. I could also have gone with version 5.59 (the
first dual-lifed version), but the CPAN testers say 5.63 (the current
version) works back to perl 5.6.1.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@3919 7ac2fd03-a204-0410-b7b1-b9f6acf132f9