RT # 67159: perl 5.12 package NAME VERSION not understood
P::C::P::Modules::RequireVersionVar now recognizes a version passed as
the second argument of a package statement.

P::C::P::ValuesAndExpressions::ProhibitMagicNumbers now allows any
number when passed as the second argument of a package statement. This
is consistent (I believe) with its handling of numbers assigned to
$VERSION.

Tests for this have been added to the relevant .run files, and (glory
be!) the Changes file has been updated.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@4062 7ac2fd03-a204-0410-b7b1-b9f6acf132f9