RT #69489 - ErrorHandling::RequireCheckingReturnValueOfEval false positive
The issue is that the subject policy does not accept grep { eval ... }.
The failing code provided in the ticket was
    grep { ! eval "require $_" } @packages;
After a bit of thought (enough, I hope!) I decided that this was the
moral equivalent of
    foreach ( @packages ) {
        eval "require $_" and next;
	...
    }
which is acceptable to this policy. So ...

This modification looks for both
    grep { eval ... } ...
and
    grep eval ... , ...
though it does not look quite as hard for the latter.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@4088 7ac2fd03-a204-0410-b7b1-b9f6acf132f9