RT #69867: RegularExpressions::ProhibitUnusedCapture false positive when capture used in an else branch
The problem was that the complainant used '!~' to bind the regexp to its
operand. The policy checks all parts of an 'if', and fell over a regular
expression in the 'if' block.

The fix was to skip the first block of an 'if' or 'elsif' if the regular
expression under analysis used '!~' to bind to the operand. Yes, this is
bogus code, and there probably should be a policy against it. But it is
not in violation of _this_ policy.

Strictly speaking the code should also skip 'if' or 'elsif' blocks other
than the first if the regexp is not bound with a '!~', but that is an
optimization (and possible false negative) that I did not tackle at this
time.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@4091 7ac2fd03-a204-0410-b7b1-b9f6acf132f9