RT #72660: Exclude pragmas from Modules::RequireExplicitPackage
This commit does not do what the original request asked for, but does
allow the user to specify modules which are allowed to be imported
before the 'package' statement. The configuration option is
'allow_import_of' (for the moment -- I'm sure someone can think of a
better name), and by default it allows nothing.


git-svn-id: http://perlcritic.tigris.org/svn/perlcritic/trunk/distributions/Perl-Critic@4098 7ac2fd03-a204-0410-b7b1-b9f6acf132f9